Add some more logic tests and fix some typos

+++ trunk/ng/src/uk/co/wilson/ng/runtime/metaclass/primitives/biginteger/GreaterThanOrEquals.java
@@ -68,7 +68,7 @@ public class GreaterThanOrEquals extends BaseBooleanComparison {
   }
 
   public boolean doApplyBoolean(final ThreadContext tc, final BigDecimal lhs, final Object rhs) {
-    return lhs.compareTo((BigDecimal)rhs) >= 0;
+    return lhs.compareTo(BigDecimal.valueOf(((BigInteger)rhs).longValue())) >= 0;
   }
 
   public boolean doApplyBoolean(final ThreadContext tc, final BigInteger lhs, final char rhs) {
